git93 add junits and securedProperties support
