+++ src/main/java/org/odpi/egeria/connectors/hms/eventmapper/HMSOMRSEventProducer.java
@@ -15,8 +15,12 @@ import org.odpi.egeria.connectors.hms.auditlog.HMSOMRSAuditCode;
 import org.odpi.egeria.connectors.hms.auditlog.HMSOMRSErrorCode;
 import org.odpi.egeria.connectors.hms.helpers.ExceptionHelper;
 import org.odpi.egeria.connectors.hms.helpers.SupportedTypes;
+import org.odpi.openmetadata.frameworks.auditlog.AuditLog;
 import org.odpi.openmetadata.frameworks.connectors.ffdc.ConnectorCheckedException;
 import org.odpi.openmetadata.frameworks.connectors.properties.EndpointProperties;
+import org.odpi.openmetadata.repositoryservices.connectors.stores.metadatacollectionstore.repositoryconnector.OMRSRepositoryConnector;
+import org.odpi.openmetadata.repositoryservices.connectors.stores.metadatacollectionstore.repositoryconnector.OMRSRepositoryHelper;
+import org.odpi.openmetadata.repositoryservices.connectors.stores.metadatacollectionstore.repositoryeventmapper.OMRSRepositoryEventProcessor;
 import org.odpi.openmetadata.repositoryservices.ffdc.exception.RepositoryErrorException;
 
 import java.util.*;
@@ -31,19 +35,29 @@ import java.util.*;
  * 1) for the asset Entities and relationships
  * 2) for each RelationalTable, it's RelationalColumns and associated relationships
  */
-public class HMSOMRSRepositoryEventMapper extends OMRSDatabasePollingRepositoryEventMapper
+public class HMSOMRSEventProducer extends OMRSEventProducer
 {
 
     private HiveMetaStoreClient client = null;
 
 
     private final String className = this.getClass().getName();
-
     /**
      * Default constructor
      */
-    public HMSOMRSRepositoryEventMapper() {
+    public HMSOMRSEventProducer() {
         super();
+    }
+
+    public HMSOMRSEventProducer(AuditLog auditLog,
+                                OMRSRepositoryHelper repositoryHelper,
+                                OMRSRepositoryConnector repositoryConnector,
+                                OMRSRepositoryEventProcessor repositoryEventProcessor,
+                                Map<String, Object> configurationProperties,
+                                EndpointProperties endpoint,
+                                String userId) {
+        super(auditLog, repositoryHelper, repositoryConnector, repositoryEventProcessor, configurationProperties, endpoint, userId);
+
     }
 
     /**
@@ -55,8 +69,6 @@ public class HMSOMRSRepositoryEventMapper extends OMRSDatabasePollingRepositoryE
     @Override
     protected void connectTo3rdParty() throws RepositoryErrorException, ConnectorCheckedException {
         String methodName = "connectTo3rdParty";
-
-        Map<String, Object> configurationProperties = connectionProperties.getConfigurationProperties();
         Boolean configuredUseSSL = (Boolean) configurationProperties.get(HMSOMRSRepositoryEventMapperProvider.USE_SSL);
         boolean useSSL = false;
         if (configuredUseSSL != null) {
@@ -72,15 +84,13 @@ public class HMSOMRSRepositoryEventMapper extends OMRSDatabasePollingRepositoryE
         if (configuredMetadataStorePassword != null) {
             metadata_store_password = configuredMetadataStorePassword;
         }
-
-        EndpointProperties endpointProperties = connectionProperties.getEndpoint();
-        if (endpointProperties == null) {
+        if (endpoint == null) {
             ExceptionHelper.raiseRepositoryErrorException(className, HMSOMRSErrorCode.ENDPOINT_NOT_SUPPLIED_IN_CONFIG, methodName, null, "null");
         } else {
             // populate the Hive configuration for the HMS client.
             Configuration conf = new Configuration();
             // we only support one thrift uri at this time
-            conf.set("metastore.thrift.uris", endpointProperties.getAddress());
+            conf.set("metastore.thrift.uris", endpoint.getAddress());
             if (useSSL) {
                 conf.set("metastore.use.SSL", "true");
                 conf.set("metastore.truststore.path", "file:///" + System.getProperty("java.home") + "/lib/security/cacerts");
@@ -106,7 +116,7 @@ public class HMSOMRSRepositoryEventMapper extends OMRSDatabasePollingRepositoryE
             }
             metadataCollection = this.repositoryConnector.getMetadataCollection();
             metadataCollectionId = metadataCollection.getMetadataCollectionId(getUserId());
-            metadataCollectionName = localServerName;
+//            metadataCollectionName = localServerName;
         }
     }
 
@@ -122,65 +132,65 @@ public class HMSOMRSRepositoryEventMapper extends OMRSDatabasePollingRepositoryE
         }
         return tableNames;
     }
-        protected ConnectorTable getTableFrom3rdParty(String catName, String dbName, String baseCanonicalName, String tableName) {
-                String  methodName = "getTableFrom3rdParty";
-                ConnectorTable connectorTable = null;
+    protected ConnectorTable getTableFrom3rdParty(String catName, String dbName, String baseCanonicalName, String tableName) {
+        String  methodName = "getTableFrom3rdParty";
+        ConnectorTable connectorTable = null;
 
-                Table hmsTable = null;
-                try {
-                    hmsTable = client.getTable(catName, dbName, tableName);
-                } catch (TException e) {
-                    auditLog.logMessage(methodName, HMSOMRSAuditCode.HIVE_GETTABLE_FAILED.getMessageDefinition(tableName, e.getMessage()));
-                }
-                if (hmsTable != null) {
-                    connectorTable = getTableFromHMSTable(baseCanonicalName, hmsTable);
-                    Iterator<FieldSchema> colsIterator = hmsTable.getSd().getColsIterator();
+        Table hmsTable = null;
+        try {
+            hmsTable = client.getTable(catName, dbName, tableName);
+        } catch (TException e) {
+            auditLog.logMessage(methodName, HMSOMRSAuditCode.HIVE_GETTABLE_FAILED.getMessageDefinition(tableName, e.getMessage()));
+        }
+        if (hmsTable != null) {
+            connectorTable = getTableFromHMSTable(baseCanonicalName, hmsTable);
+            Iterator<FieldSchema> colsIterator = hmsTable.getSd().getColsIterator();
 
-                    while (colsIterator.hasNext()) {
-                        FieldSchema fieldSchema = colsIterator.next();
-                        String columnName = fieldSchema.getName();
-                        String dataType = fieldSchema.getType();
+            while (colsIterator.hasNext()) {
+                FieldSchema fieldSchema = colsIterator.next();
+                String columnName = fieldSchema.getName();
+                String dataType = fieldSchema.getType();
 
-                        ConnectorColumn column = new ConnectorColumn();
-                        column.setName(columnName);
-                        column.setQualifiedName(connectorTable.getQualifiedName() + SupportedTypes.SEPARATOR_CHAR + columnName);
-                        column.setType(dataType);
-                        connectorTable.addColumn(column);
-                    }
+                ConnectorColumn column = new ConnectorColumn();
+                column.setName(columnName);
+                column.setQualifiedName(connectorTable.getQualifiedName() + SupportedTypes.SEPARATOR_CHAR + columnName);
+                column.setType(dataType);
+                connectorTable.addColumn(column);
+            }
         }
         return connectorTable;
     }
 
 
-        private ConnectorTable getTableFromHMSTable(String baseCanonicalName, Table hmsTable) {
-            ConnectorTable connectorTable = new ConnectorTable();
-            String name = hmsTable.getTableName();
-            String tableType = hmsTable.getTableType();
-            String tableCanonicalName = baseCanonicalName + SupportedTypes.SEPARATOR_CHAR + "schema" + SupportedTypes.SEPARATOR_CHAR + name;
-            String typeName = SupportedTypes.TABLE;
-            int createTime = hmsTable.getCreateTime();
-            //                            String owner = hmsTable.getOwner();
-            //                            if (owner != null) {
-            //                               TODO Can we store this on the hmsTable ?
-            //                            }
+    private ConnectorTable getTableFromHMSTable(String baseCanonicalName, Table hmsTable) {
+        ConnectorTable connectorTable = new ConnectorTable();
+        String name = hmsTable.getTableName();
+        String tableType = hmsTable.getTableType();
+        String tableCanonicalName = baseCanonicalName + SupportedTypes.SEPARATOR_CHAR + SupportedTypes.SCHEMA_TOKEN_NAME + SupportedTypes.SEPARATOR_CHAR + name;
+        String typeName = SupportedTypes.TABLE;
+        int createTime = hmsTable.getCreateTime();
+        //                            String owner = hmsTable.getOwner();
+        //                            if (owner != null) {
+        //                               TODO Can we store this on the hmsTable ?
+        //                            }
 
-            connectorTable.setName(name);
-            connectorTable.setCreateTime(new Date(createTime));
-            connectorTable.setQualifiedName(tableCanonicalName);
-            connectorTable.setType(tableType);
-            connectorTable.setType(typeName);
+        connectorTable.setName(name);
+        connectorTable.setCreateTime(new Date(createTime));
+        connectorTable.setQualifiedName(tableCanonicalName);
+        connectorTable.setType(tableType);
+        connectorTable.setType(typeName);
 
-            Iterator<FieldSchema> colsIterator = hmsTable.getSd().getColsIterator();
+        Iterator<FieldSchema> colsIterator = hmsTable.getSd().getColsIterator();
 
-            while (colsIterator.hasNext()) {
-                ConnectorColumn column = new ConnectorColumn();
-                FieldSchema fieldSchema = colsIterator.next();
-                String columnName = fieldSchema.getName();
-                column.setName(columnName);
-                column.setType(fieldSchema.getType());
-                column.setQualifiedName(tableCanonicalName + SupportedTypes.SEPARATOR_CHAR + columnName);
-            }
-
-            return connectorTable;
+        while (colsIterator.hasNext()) {
+            ConnectorColumn column = new ConnectorColumn();
+            FieldSchema fieldSchema = colsIterator.next();
+            String columnName = fieldSchema.getName();
+            column.setName(columnName);
+            column.setType(fieldSchema.getType());
+            column.setQualifiedName(tableCanonicalName + SupportedTypes.SEPARATOR_CHAR + columnName);
         }
+
+        return connectorTable;
+    }
 }
