+++ src/main/java/org/odpi/egeria/connectors/hms/eventmapper/HMSOMRSRepositoryEventMapperProvider.java
@@ -38,6 +38,8 @@ public class HMSOMRSRepositoryEventMapperProvider extends OMRSRepositoryConnecto
 
     static final String CACHE_INTO_CACHING_REPOSITORY = "cacheIntoCachingRepository";
 
+    static final String CONNECTION_SECURED_PROPERTIES = "connectionSecuredProperties";
+
     /**
      * If this is set then we use this as the endpoint address (e.g. the JDBC URL)
      * If it is not set then, no connection is associated with the asset
@@ -49,7 +51,7 @@ public class HMSOMRSRepositoryEventMapperProvider extends OMRSRepositoryConnecto
      * OMRS Connector implementation.
      */
     public HMSOMRSRepositoryEventMapperProvider() {
-        Class<?> connectorClass = HMSOMRSRepositoryEventMapper.class;
+        Class<?> connectorClass = OMRSDatabasePollingRepositoryEventMapper.class;
         super.setConnectorClassName(connectorClass.getName());
         ConnectorType connectorType = new ConnectorType();
         connectorType.setType(ConnectorType.getConnectorTypeType());
@@ -70,6 +72,7 @@ public class HMSOMRSRepositoryEventMapperProvider extends OMRSRepositoryConnecto
         knownConfigProperties.add(METADATA_STORE_USER);
         knownConfigProperties.add(METADATA_STORE_PASSWORD);
         knownConfigProperties.add(USE_SSL);
+        knownConfigProperties.add(CONNECTION_SECURED_PROPERTIES);
 
         connectorType.setRecognizedConfigurationProperties(knownConfigProperties);
