+++ src/test/resources/securedProperties-bad-as-str.json
@@ -1,0 +1,1 @@
+aaaa
