+++ src/test/resources/securedProperties.json
@@ -1,0 +1,5 @@
+{
+  "aaa": "aaa-value",
+  "bbb": "bbb-value",
+  "ccc": "ccc-value"
+}
